4.DevConnect/devconnect/requirements.txt
# Django and Core Dependencies
Django>=4.2.0,<5.0.0
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.2.0
django-cors-headers>=4.0.0
django-filter>=23.0
django-decouple>=3.8

# Database and ORM
psycopg2-binary>=2.9.0  # PostgreSQL adapter
mysqlclient>=2.1.0      # MySQL adapter (alternative)

# Real-time Communication
channels>=4.0.0
channels-redis>=4.1.0
daphne>=4.0.0

# Background Tasks
celery>=5.3.0
redis>=4.5.0
django-celery-beat>=2.5.0

# Image Processing
Pillow>=10.0.0

# HTTP and API
requests>=2.31.0
urllib3>=2.0.0

# Development and Testing
pytest>=7.4.0
pytest-django>=4.5.0
coverage>=7.2.0
factory-boy>=3.2.0

# Code Quality
black>=23.0.0
flake8>=6.0.0
isort>=5.12.0

# Environment and Configuration
python-decouple>=3.8
python-dotenv>=1.0.0

# Internationalization
Babel>=2.12.0

# Security
cryptography>=41.0.0

# Monitoring and Logging
sentry-sdk>=1.28.0

# Production
gunicorn>=21.0.0
whitenoise>=6.5.0
